+++ EscolaDBWinForm/Models/IAnoLetivo.cs
@@ -10,7 +10,7 @@ namespace EscolaDBWinForm.Models
     {
         void Add(AnoLetivo ano);
         void Edit(AnoLetivo ano);
-        void Delete(int id);
+        void Delete(short id);
         IEnumerable<AnoLetivo> GetAll();
         IEnumerable<AnoLetivo> GetByValue(string value); //Search por valor
         ICollection<Inscricao> GetInscricao(int id); //Busca todas as Inscricoes de um Ano para apresentar na gridview
