+++ EscolaDBWinForm/Models/IEstadoEpoca.cs
@@ -11,7 +11,7 @@ namespace EscolaDBWinForm.Models
         //Layer de acesso a dados CRUD
         void Add(EstadoEpoca e_epoca);
         void Edit(EstadoEpoca e_espoca);
-        void Delete(int id);
+        void Delete(short id);
         IEnumerable<EstadoEpoca> GetAll();
         IEnumerable<EstadoEpoca> GetByValue(string value); //Search por valor
         ICollection<Inscricao> GetInscricao(int id); //Busca todas as Inscricoes de uma Epoca para apresentar na gridview
